+++ src/DnnPackager/Command/CommandVisitor.cs
@@ -19,7 +19,6 @@ namespace DnnPackager.Command
 
         public bool Success { get; set; }
 
-
         public void VisitBuildCommand(BuildOptions options)
         {
             FileInfo[] installPackages = null;
@@ -90,82 +89,10 @@ namespace DnnPackager.Command
             Microsoft.Build.Evaluation.ProjectCollection collection = new Microsoft.Build.Evaluation.ProjectCollection();
             Microsoft.Build.Evaluation.Project project = new Microsoft.Build.Evaluation.Project(options.ProjectName, null, null, collection, Microsoft.Build.Evaluation.ProjectLoadSettings.IgnoreMissingImports);
 
-            //  project.Load(options.ProjectName, ProjectLoadSettings.IgnoreMissingImports);
-
-            var projectDir = Path.GetDirectoryName(options.ProjectName);
-            _Logger.LogInfo(string.Format("Project Dir is: {0}", projectDir));
-
-            var globalPropsFileName = "DnnPackager.props";
+            var installHelper = new InstallTargetsHelper(_Logger);
             var toolsDir = options.ToolsPath.TrimEnd(new char[] { '\\', '/' });
-
-            var globalPropsFilePath = Path.Combine(toolsDir, globalPropsFileName);
-
-            EnsureImport(project, globalPropsFilePath);
-
-            var projectPropsFileName = "DnnPackageBuilderOverrides.props";
-            var projectPropsFilePath = Path.Combine(projectDir, projectPropsFileName);
-
-            EnsureImport(project, projectPropsFilePath);
-
-            var targetsFileName = "dnnpackager.targets";
-            var targetsFilePath = Path.Combine(toolsDir, targetsFileName);
-
-            EnsureImport(project, targetsFilePath);
-
-            // remove legacy imports.
-            RemoveImport(project, "DnnPackager.Build.targets");
-
-            // if octopack targets are there ensure they are added after other targets.
-            ReImportTargetIfExists(project, "OctoPack.targets");
-
-            project.Save(options.ProjectName);
-            Success = true;
-        }
-
-        private void ReImportTargetIfExists(Microsoft.Build.Evaluation.Project project, string importProjectPath)
-        {
-            var fileName = Path.GetFileName(importProjectPath);
-            var existingImport = project.Xml.Imports.Cast<ProjectImportElement>().FirstOrDefault(i => i.Project.EndsWith(fileName));
-
-            if (existingImport != null)
-            {
-                project.Xml.RemoveChild(existingImport);
-                project.Xml.AddImport(existingImport.Project);              
-                _Logger.LogInfo(string.Format("Re-imported: {0}", fileName));
-            }
-
-        }
-
-        private void RemoveImport(Microsoft.Build.Evaluation.Project project, string importProjectPath)
-        {
-            var fileName = Path.GetFileName(importProjectPath);
-            var existingImport = project.Xml.Imports.Cast<ProjectImportElement>().FirstOrDefault(i => i.Project.EndsWith(fileName));           
-            if (existingImport != null)
-            {
-                project.Xml.RemoveChild(existingImport);            
-                _Logger.LogInfo(string.Format("Removed import of: {0}", fileName));
-            }
-        }
-
-        private void EnsureImport(Microsoft.Build.Evaluation.Project project, string importProjectPath)
-        {
-            // Ensure import is present, replace existing if found.
-            var fileName = Path.GetFileName(importProjectPath);
-            var existingImport = project.Xml.Imports.Cast<ProjectImportElement>().FirstOrDefault(i => i.Project.EndsWith(fileName));          
-            if (existingImport != null)
-            {
-                _Logger.LogInfo(string.Format("The existing import will be upgraded for: {0}", fileName));
-                project.Xml.RemoveChild(existingImport);           
-
-            }
-
-            var projectUri = new Uri(string.Format("file://{0}", project.FullPath));
-            var importFileUri = new Uri(string.Format("file://{0}", importProjectPath));
-
-            var relativeImportFile = projectUri.MakeRelativeUri(importFileUri).ToString().Replace('/', '\\');
-            project.Xml.AddImport(relativeImportFile);            
-         
-            _Logger.LogInfo(string.Format("Successfully imported: {0}", relativeImportFile));
+            this.Success = installHelper.Install(project, toolsDir);      
+          
         }
 
         public bool BuildProjectAndGetOutputZips(BuildOptions options, out FileInfo[] installPackages, out DTE dte)
